Move geospatial headers to subfolder
This simplifies the logic required for packaging.
