+++ geospatial/include/ignition/common/geospatial/ImageHeightmap.hh
@@ -14,16 +14,16 @@
  * limitations under the License.
  *
 */
-#ifndef IGNITION_COMMON_IMAGEHEIGHTMAPDATA_HH_
-#define IGNITION_COMMON_IMAGEHEIGHTMAPDATA_HH_
+#ifndef IGNITION_COMMON_GEOSPATIAL_IMAGEHEIGHTMAPDATA_HH_
+#define IGNITION_COMMON_GEOSPATIAL_IMAGEHEIGHTMAPDATA_HH_
 
 #include <limits>
 #include <string>
 #include <vector>
 #include <ignition/math/Vector3.hh>
 
-#include <ignition/common/graphics/Export.hh>
-#include <ignition/common/HeightmapData.hh>
+#include <ignition/common/geospatial/Export.hh>
+#include <ignition/common/geospatial/HeightmapData.hh>
 #include <ignition/common/Image.hh>
 
 namespace ignition
@@ -31,7 +31,7 @@ namespace ignition
   namespace common
   {
     /// \brief Encapsulates an image that will be interpreted as a heightmap.
-    class IGNITION_COMMON_GRAPHICS_VISIBLE ImageHeightmap
+    class IGNITION_COMMON_GEOSPATIAL_VISIBLE ImageHeightmap
       : public ignition::common::HeightmapData
     {
       /// \brief Constructor
